+++ src/main/java/subway/line/Line.java
@@ -1,24 +1,33 @@
-package subway;
+package subway.line;
 
+import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import subway.station.Station;
 
 @Entity
-public class Station {
+public class Line {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(length = 20, nullable = false)
     private String name;
+    private String color;
+    @ManyToMany
+    private List<Station> stations;
 
-    public Station() {
+    public Line() {
     }
 
-    public Station(String name) {
+    public Line(String name, String color, List<Station> stations) {
         this.name = name;
+        this.color = color;
+        this.stations = stations;
     }
 
     public Long getId() {
@@ -29,7 +38,11 @@ public class Station {
         return name;
     }
 
-    public void update(String name) {
-        this.name = name;
+    public String getColor() {
+        return color;
+    }
+
+    public List<Station> getStations() {
+        return stations;
     }
 }
