+++ src/main/java/subway/line/LineController.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.line;
 
 import java.net.URI;
 import java.util.List;
@@ -7,6 +7,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import subway.station.Station;
+import subway.station.StationService;
 
 @RestController
 public class LineController {
