+++ src/main/java/subway/line/LineRequest.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.line;
 
 import java.util.List;
+import subway.station.Station;
 
 public class LineRequest {
