+++ src/main/java/subway/line/LineService.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.line;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
