+++ src/main/java/subway/station/Station.java
@@ -1,32 +1,24 @@
-package subway;
+package subway.station;
 
-import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToMany;
 
 @Entity
-public class Line {
-
+public class Station {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(length = 20, nullable = false)
     private String name;
-    private String color;
-    @ManyToMany
-    private List<Station> stations;
 
-    public Line() {
+    public Station() {
     }
 
-    public Line(String name, String color, List<Station> stations) {
+    public Station(String name) {
         this.name = name;
-        this.color = color;
-        this.stations = stations;
     }
 
     public Long getId() {
@@ -37,11 +29,7 @@ public class Line {
         return name;
     }
 
-    public String getColor() {
-        return color;
-    }
-
-    public List<Station> getStations() {
-        return stations;
+    public void update(String name) {
+        this.name = name;
     }
 }
